fix: Support IP logging behind a reverse proxy
Previously, this would always log the proxy's internal address, if behind
a reverse proxy, or the real remote address if not. By combining the
apparent remote_addr with the XFF chain, we can get the best possible
information without having to know whether we're behind a proxy. (This
assumes the proxy is set up to append to the XFF chain, so the example in
the README is updated and corrected.)